Move purge_trash method to DossierResolver class.
